pyproject.toml
[project]
name = "hearts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "edfio>=0.4.9",
    "json-repair>=0.51.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "loguru>=0.7.3",
    "mne>=1.10.1",
    "neurokit2>=0.2.12",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "rich>=14.1.0",
    "yasa>=0.6.5",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "tornado>=6.5.2",
    "pyzmq>=27.1.0",
    "jupyter-client>=8.6.3",
    "pyperclip>=1.11.0",
    "openpyxl>=3.1.5",
    "xlrd>=2.0.2",
    "pyyaml>=6.0.3",
    "xgboost>=3.1.1",
    "statsmodels>=0.14.5",
    "shap>=0.49.1",
    "vitaldb>=1.5.8",
    "plotly>=6.4.0",
    "heartpy>=1.2.7",
    "py-ecg-detectors>=1.3.5",
    "biosppy>=2.2.4",
    "hrv-analysis>=1.0.5",
    "pysiology>=0.0.9.6",
    "kaggle>=1.7.4.5",
    "typer>=0.20.0",
    "langchain-aws>=1.1.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "datasets>=3.0.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "soundfile>=0.12.0",
    "pytest>=9.0.2",
    "librosa>=0.11.0",
    "pyarrow>=22.0.0",
    "langchain-google-genai>=4.1.2",
    "claude-agent-sdk>=0.1.18",
    "langchain-google-vertexai>=3.2.1",
    "tabulate>=0.9.0",
    # biomni dependencies
    "mcp>=1.25.0",
    "gradio>=6.3.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "networkx>=3.6.1",
    "requests>=2.32.5",
    "notebook>=7.5.0",
    "sentencepiece>=0.2.1",
    "tooluniverse>=1.0.15.2",
    "python-dotenv>=1.2.1",
    "langchain-ollama>=1.0.1",
    "biomni",
    "langchain-xai>=1.1.0",
]

# Optional dependencies for specific use cases
# Uncomment if you need to preprocess raw Bridge2AI Voice audio files
# Note: Raw audio requires credentialed access via DACO@b2ai-voice.org
# The NCSA dataset already contains preprocessed features, so this is typically not needed
# "b2aiprep>=0.1.0",  # Bridge2AI Voice preprocessing tools
# "senselab>=0.1.0",  # Audio processing utilities used by b2aiprep

[tool.pyright]
# Paths to completely ignore during type checking / indexing
exclude = [
    ".venv/**",
    "data/**"
]
useLibraryCodeForTypes = true
typeCheckingMode = "off"

[tool.uv.sources]
biomni = { git = "https://github.com/yang-ai-lab/Biomni.git", branch = "main" }

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
